// graphql/generated/schematypes.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
	ID: string;
	String: string;
	Boolean: boolean;
	Int: number;
	Float: number;
};
// ...
export type ContractFunctionInputType = {
	__typename?: 'ContractFunctionInputType';
	internalType?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	type?: Maybe<Scalars['String']>;
};

export type ContractFunctionInputInputType = {
	internalType?: InputMaybe<Scalars['String']>;
	name?: InputMaybe<Scalars['String']>;
	indexed?: InputMaybe<Scalars['Boolean']>;
	anonymous?: InputMaybe<Scalars['Boolean']>;
	type?: InputMaybe<Scalars['String']>;
};

export type ContractAbiElement = {
	__typename?: 'ContractAbiElement';
	inputs?: Maybe<Array<Maybe<ContractFunctionInputType>>>;
	name?: Maybe<Scalars['String']>;
	outputs?: Maybe<Array<Maybe<ContractFunctionInputType>>>;
	stateMutability?: Maybe<Scalars['String']>;
	type?: Maybe<Scalars['String']>;
};
// ...
export type ReturnValues = {
	__typename?: 'ReturnValues';
	from?: Maybe<Scalars['String']>;
	to?: Maybe<Scalars['String']>;
	num?: Maybe<Scalars['String']>;
};

export type Raw = {
	__typename?: 'Raw';
	data?: Maybe<Scalars['String']>;
	topics?: Maybe<Array<Maybe<Scalars['String']>>>;
};

export type ContractEvent = {
	__typename?: 'ContractEvent';
	contractString?: Maybe<Scalars['String']>;
	contractAbi?: Maybe<Array<Maybe<ContractAbiElement>>>;
	eventName?: Maybe<Scalars['String']>;
	fromHeight?: Maybe<Scalars['Int']>;
	toHeight?: Maybe<Scalars['Int']>;
	returnValues?: Maybe<ReturnValues>;
	event?: Maybe<Scalars['String']>;
	raw?: Maybe<Raw>;
	accountBlockHeight?: Maybe<Scalars['String']>;
	accountBlockHash?: Maybe<Scalars['String']>;
	contractAddress?: Maybe<Scalars['String']>;
	logAddress?: Maybe<Scalars['String']>;
};

export type TokenInfo = {
	__typename?: 'TokenInfo';
	name?: Maybe<Scalars['String']>;
	contractAddress?: Maybe<Scalars['String']>;
	totalSupply?: Maybe<Scalars['String']>;
	decimals?: Maybe<Scalars['Int']>;
	owner?: Maybe<Scalars['String']>;
	tokenId?: Maybe<Scalars['String']>;
	maxSupply?: Maybe<Scalars['String']>;
	index?: Maybe<Scalars['Int']>;
};

export type TokenInfoAccount = {
	__typename?: 'TokenInfoAccount';
	balance?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	symbol?: Maybe<Scalars['String']>;
	totalSupply?: Maybe<Scalars['String']>;
	decimals?: Maybe<Scalars['Int']>;
	owner?: Maybe<Scalars['String']>;
	tokenId?: Maybe<Scalars['String']>;
	maxSupply?: Maybe<Scalars['String']>;
	index?: Maybe<Scalars['Int']>;
};

export type Account = {
	__typename?: 'Account';
	address?: Maybe<Scalars['String']>;
	balances?: Maybe<Array<Maybe<TokenInfoAccount>>>;
};